requirements.txt
pymongo>=4.0.0
python-dotenv>=1.0.0
cloudscraper>=1.2.71
tqdm>=4.66.0
flask>=3.0.0
flask-cors>=4.0.0
ortools>=9.8.0
